🚧🔧🥒 reduce capybara config file
