+++ features/step_definitions/cms/page_template_steps.rb
@@ -9,7 +9,7 @@ When "I fill in the draft with {string}" do |text|
 end
 
 def fill_draft(text)
-  raise 'Please mark this scenario with @javascript if you want to work with codemirror.' unless @javascript
+  raise 'Please mark this scenario with @javascript if you want to work with codemirror.' unless javascript_test?
 
   find('#cms_template_draft', visible: :all)
   execute_script("$('#cms_template_draft').data('codemirror').setValue(#{text.inspect});")
