+++ features/step_definitions/web_steps.rb
@@ -143,14 +143,10 @@ end
 
 Then /^(.+) and confirm the dialog(?: "(.*)")?$/ do |original, text|
   ActiveSupport::Deprecation.warn "🥒 Replace with step 'And confirm the dialog'"
-  if rack_test?
+  accept_confirm(text) do
     step original
-  else
-    accept_confirm(text) do
-      step original
-    end
-    wait_for_requests
   end
+  wait_for_requests
 end
 
 Then "(they )should see the following details(:)" do |table|
