+++ features/support/capybara.rb
@@ -1,117 +1,49 @@
 # frozen_string_literal: true
 
 require 'selenium/webdriver'
-require 'capybara/minitest'
-include Capybara::Minitest::Assertions
 
-# in case firefox is needed!
-#Capybara.register_driver :selenium do |app|
-#  Capybara::Selenium::Driver.new(app, :browser => :firefox)
-#end
-
-DEFAULT_JS_DRIVER = :headless_chrome
-# in case firefox is needed!
-# DEFAULT_JS_DRIVER = :headless_firefox
-
-Capybara.default_driver = :rack_test
-Capybara.javascript_driver = DEFAULT_JS_DRIVER
-Capybara.default_selector    = :css
+# Capybara.default_max_wait_time = 10
 Capybara.disable_animation = true
+Capybara.javascript_driver = :headless_chrome
+Capybara.match = :prefer_exact # TODO: this is Capybara v1 default, we should use :one or :smart
+Capybara.server = :webrick # default is :puma
 
-# Capybara 3 changes the default server to Puma. It can be reverted to the previous default of WEBRick by specifying:
-Capybara.server = :webrick
-
-# see http://www.elabs.se/blog/60-introducing-capybara-2-1
-Capybara.configure do |config|
-  config.default_driver = :rack_test
-  config.javascript_driver = DEFAULT_JS_DRIVER
-  config.raise_server_errors = true
-  config.match = :prefer_exact
-  config.always_include_port = true
-  config.default_max_wait_time = 10
-end
-
-# Needed because cucumber-rails requires capybara/cucumber
-# https://github.com/cucumber/cucumber-rails/blob/7b47bf1dda3368247bf2d45bcb17a224e80ec6fd/lib/cucumber/rails/capybara.rb#L3
-# https://github.com/teamcapybara/capybara/blob/2.18.0/lib/capybara/cucumber.rb#L17-L19
-Before '@javascript' do
-  Capybara.current_driver = DEFAULT_JS_DRIVER
-end
+BASE_DRIVER_OPTIONS = {
+  args: [
+    # This is added in Capybara's default drivers. See https://github.com/teamcapybara/capybara/blob/0480f90168a40780d1398c75031a255c1819dce8/lib/capybara/registrations/drivers.rb#L37-L38
+    # Workaround https://bugs.chromium.org/p/chromedriver/issues/detail?id=2650&q=load&sort=-id&colspec=ID%20Status%20Pri%20Owner%20Summary
+    '--disable-site-isolation-trials',
+    # When width < 1200px, vertical navigation overlaps the page's main content and that will make
+    # some cucumbers fail
+    '--window-size=1200,2048',
+  ]
+}.freeze
 
 Before '@chrome' do
   Capybara.current_driver = :chrome
 end
 
-Before '@firefox' do
-  Capybara.current_driver = :firefox
-end
-
-Around '@security' do |scenario, block|
-  with_forgery_protection(&block)
-end
-
-# monkeypatch to fix
-# not opened for reading (IOError)
-# /cucumber-1.3.20/lib/cucumber/formatter/interceptor.rb:33:in `each'
-# /cucumber-1.3.20/lib/cucumber/formatter/interceptor.rb:33:in `collect'
-# /cucumber-1.3.20/lib/cucumber/formatter/interceptor.rb:33:in `method_missing'
-require 'cucumber/formatter/interceptor'
-class Cucumber::Formatter::Interceptor::Pipe
-  def is_a?(klass)
-    super || klass == IO
-  end
-end
-
-Capybara.register_driver :firefox do |app|
-  Capybara::Selenium::Driver.new(app, browser: :firefox)
-end
-
-Capybara.register_driver :headless_firefox do |app|
-  options = Selenium::WebDriver::Firefox::Options.new
-
-  options.add_argument('-headless')
-  options.add_argument('--window-size=1280,2048')
-
-  driver = Capybara::Selenium::Driver.new(app, browser: :firefox, options: options)
-
-  driver
-end
-
-Capybara.register_driver :firefox do |app|
-  options = Selenium::WebDriver::Firefox::Options.new
-
-  options.add_argument('--window-size=1280,2048')
-
-  driver = Capybara::Selenium::Driver.new(app, browser: :firefox, options: options)
-
-  driver
-end
-
+# Use this driver to debug scenarios locally by adding tag @chrome on top of the scenario
 Capybara.register_driver :chrome do |app|
-  options = Selenium::WebDriver::Chrome::Options.new
-  options.add_argument('--window-size=1280,2048')
-  options.add_argument('--disable-search-engine-choice-screen')
+  options = Selenium::WebDriver::Options.chrome(**BASE_DRIVER_OPTIONS)
+
   Capybara::Selenium::Driver.new(app, browser: :chrome, options: options)
 end
 
 Capybara.register_driver :headless_chrome do |app|
-  options = Selenium::WebDriver::Options.chrome(
-    logging_prefs: { performance: 'ALL', browser: 'ALL' },
-    perf_logging_prefs: { enableNetwork: true }
-  )
+  options = Selenium::WebDriver::Options.chrome(**BASE_DRIVER_OPTIONS)
+
+  options.logging_prefs = { performance: 'ALL', browser: 'ALL' }
+
+  options.add_option(:perf_logging_prefs, enableNetwork: true)
 
   options.add_argument('--headless=new')
   options.add_argument('--no-sandbox')
   options.add_argument('--disable-popup-blocking')
-  options.add_argument('--window-size=1280,2048')
   options.add_argument('--host-resolver-rules=MAP * ~NOTFOUND , EXCLUDE *localhost*')
-  options.add_argument('--disable-search-engine-choice-screen')
   options.add_argument('--disable-gpu')
 
   options.add_preference(:browser, set_download_behavior: { behavior: 'allow' })
 
-  timeout = 120 # default 60
-  client = Selenium::WebDriver::Remote::Http::Default.new(open_timeout: timeout, read_timeout: timeout)
-
-  Capybara::Selenium::Driver.new(app, browser: :chrome, options: options, http_client: client)
+  Capybara::Selenium::Driver.new(app, browser: :chrome, options: options, timeout: 120)
 end
