+++ features/support/helpers/requests_helper.rb
@@ -103,10 +103,6 @@ module WaitForRequests
 
     Capybara.page.evaluate_script('jQuery.active').zero?
   end
-
-  def javascript_test?
-    Capybara.current_driver == Capybara.javascript_driver
-  end
 end
